Fix for x86_64.
	* src/engine/QueryCore.c: Fix for x86_64.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/1.8@10280 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/QueryCore.c
@@ -669,7 +669,7 @@ static char * int64_to_string (gpointer object, QueryAccess get)
 {
   gint64 num = ((query_int64_getter)get)(object);
 
-  return g_strdup_printf (GNC_SCANF_LLD, num);
+  return g_strdup_printf (GNC_SCANF_LLD, (long long int) num);
 }
 
 /* QUERYCORE_DOUBLE */
